move rollup methods to mixin to test it
and

pass time_range parameter
